+++ apps/app/components/command-palette/index.ts
@@ -1,5 +1,6 @@
 export * from "./issue";
 export * from "./change-interface-theme";
+export * from "./command-k";
 export * from "./command-pallette";
 export * from "./helpers";
 export * from "./shortcuts-modal";
